Add my vimstuff repo as a subtree.
Refer to https://github.com/WiserYaldabaoth/vimstuff

It's more awkward than I thought it would be to move from a .vim
repository to a dotfiles repository, but I think I have it working.
The only problem I foresee with this setup is that it will make it more
difficult to change my .vim settings "on the fly". I have my .vim repo
scattered throughout different systems, so I think I'll be able to
figure it out.