Allow for sourcing local .tmux.conf file.
There are some settings that might need to be on some machines that
other machines won't need -- for example, in cygwin, I need tmux to see
the terminal as an xterm-256color terminal. In order to maximize
portability, I can move that to a .tmux.conf.local file and tell tmux to
source that at the beginning!