Update .tmux.conf to match my own theme more.
I tweaked the statusbar to match gruvbox a little closer, e.g. modifying
the windows indicators in the status bar to look like more like the
gruvbox airline tabs in vim. The unecessary guillemets are gone until I
can figure out how to conditionally display powerline-font triangles
based on their presence in the current font (if that is even possible).